tushare_to_clickhouse/pyproject.toml
[project]
name = "tushare-to-clickhouse"
version = "0.1.0"
description = "Sync Tushare data to ClickHouse"
requires-python = ">=3.10"
dependencies = [
    "tushare>=1.3.0",
    "clickhouse-connect>=0.7.0",
    "pandas>=2.0.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
]

[project.scripts]
tushare-to-clickhouse = "tushare_to_clickhouse.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
